Crate ticketed_lock

Ticketed lock.

Ticketed lock is similar to RwLock, except that the acquisition of the lock is split:

obtaining a ticket, which has to be done on the same thread as the locked storage

waiting on a ticket, which puts the current thread to sleep until the ticket is due. That moment comes when all the previous tickets are processed.

working with the data behind a read/lock guard

when the guard is freed, it allows the following tickets to become active

A ticket can be moved between threads or even just lost. Consecutive read-only tickets do not guarantee a particular lock order. All the ticket counting is done based on Arc primitives, and the only unsafe code that this library has is for accessing the actual data behind a guard.

Structs

ReadLockGuard
The read-only guard of data, allowing &T dereferences.

ReadTicket
A ticket to read the data at some point.

TicketedLock
The ticketed lock, which wraps the data.

WriteLockGuard
The read-only guard of data, allowing &T dereferences.

WriteTicket
A ticket to read/write the data at some point.